graphDirective to landingController transfer
I essentially copy-pasted all of the graphDirective controller to the
landingController. I then isolated the scope of graphDirective to take
in the "allWunderData" json that now the landingController makes.  I
then used the $scope.$watch() method to asynchronously take in data from
the landingController so that the cssColors service could lazy load into
the graphDirective html. Finally I refactored everything. went from 200+
lines of code to just 50 (and less console logs).